Refined FinancialStatement.py
